# demo/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Admin API
  description: API для управления банковскими картами администратором
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
  - url: https://api.example.com
    description: Продакшен сервер
paths:
...
  /user/card/balance:
    get:
      summary: Получить баланс карты
      description: Возвращает баланс указанной карты пользователя
      operationId: getCardBalance
      tags:
        - User Card Management
      parameters:
        - name: Authorization
          in: header
          description: Bearer токен авторизации
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: cardId
          in: query
          description: ID карты
          required: true
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: number
                format: decimal
                example: 1500.75
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card:
    get:
      summary: Получить список карт пользователя
      description: Возвращает пагинированный список банковских карт текущего пользователя
      operationId: getUsersCards
      tags:
        - User Card Management
      parameters:
        - name: page
          in: query
          description: Номер страницы (начинается с 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sortBy
          in: query
          description: Поле для сортировки
          required: false
          schema:
            type: string
            default: "id"
        - name: sortDirection
          in: query
          description: Направление сортировки
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: "asc"
        - name: Authorization
          in: header
          description: Bearer токен авторизации
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardDTO'
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Создать новую карту
      description: Создает новую банковскую карту для текущего пользователя
      operationId: createCard
      tags:
        - User Card Management
      parameters:
        - name: Authorization
          in: header
          description: Bearer токен авторизации
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card/active:
    get:
      summary: Получить список активных карт пользователя
      description: Возвращает пагинированный список активных банковских карт текущего пользователя
      operationId: getActiveUsersCards
      tags:
        - User Card Management
      parameters:
        - name: page
          in: query
          description: Номер страницы (начинается с 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sortBy
          in: query
          description: Поле для сортировки
          required: false
          schema:
            type: string
            default: "id"
        - name: sortDirection
          in: query
          description: Направление сортировки
          required: false
          schema:
            type: string
            enum: [ asc, desc ]
            default: "asc"
        - name: Authorization
          in: header
          description: Bearer токен авторизации
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardDTO'
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card/blocked:
    get:
      summary: Получить список заблокированных карт пользователя
      description: Возвращает пагинированный список заблокированных банковских карт текущего пользователя
      operationId: getBlockedUsersCards
      tags:
        - User Card Management
      parameters:
        - name: page
          in: query
          description: Номер страницы (начинается с 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sortBy
          in: query
          description: Поле для сортировки
          required: false
          schema:
            type: string
            default: "id"
        - name: sortDirection
          in: query
          description: Направление сортировки
          required: false
          schema:
            type: string
            enum: [ asc, desc ]
            default: "asc"
        - name: Authorization
          in: header
          description: Bearer токен авторизации
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardDTO'
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card/expired:
    get:
      summary: Получить список просроченных карт пользователя
      description: Возвращает пагинированный список просроченных банковских карт текущего пользователя
      operationId: getExpiredUsersCards
      tags:
        - User Card Management
      parameters:
        - name: page
          in: query
          description: Номер страницы (начинается с 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sortBy
          in: query
          description: Поле для сортировки
          required: false
          schema:
            type: string
            default: "id"
        - name: sortDirection
          in: query
          description: Направление сортировки
          required: false
          schema:
            type: string
            enum: [ asc, desc ]
            default: "asc"
        - name: Authorization
          in: header
          description: Bearer токен авторизации
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardDTO'
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card/block:
    put:
      summary: Запрос на блокировку карты
      description: Отправляет запрос на блокировку карты администратору
      operationId: blockCardRequest
      tags:
        - User Card Management
      parameters:
        - name: Authorization
          in: header
          description: Bearer токен авторизации
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: cardId
          in: query
          description: ID карты для блокировки
          required: true
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        '200':
          description: Запрос на блокировку отправлен
          content:
            text/plain:
              schema:
                type: string
                example: "Block request from user-id:123 has been sent to admin.\n Card to block: 12345"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/token:
    post:
      summary: Получить токен авторизации
      description: Возвращает JWT токен для авторизации пользователя
      operationId: getToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreationDTO'
      responses:
        '200':
          description: Успешная авторизация
          content:
            text/plain:
              schema:
                type: string
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Неверные учетные данные
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card/deposit:
    put:
      summary: Пополнить карту
      description: Пополняет баланс указанной карты
      operationId: deposit
      tags:
        - User Card Management
      parameters:
        - name: amount
          in: query
          description: Сумма для пополнения
          required: true
          schema:
            type: number
            format: decimal
            minimum: 0.01
            example: 1000.00
        - name: cardId
          in: query
          description: ID карты для пополнения
          required: true
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        '200':
          description: Средства успешно зачислены
          content:
            text/plain:
              schema:
                type: string
                example: "Funds have been deposited: +1000.00"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card/withdraw:
    put:
      summary: Снять средства с карты
      description: Снимает средства с указанной карты
      operationId: withdraw
      tags:
        - User Card Management
      parameters:
        - name: amount
          in: query
          description: Сумма для снятия
          required: true
          schema:
            type: number
            format: decimal
            minimum: 0.01
            example: 500.00
        - name: cardId
          in: query
          description: ID карты для снятия
          required: true
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        '200':
          description: Средства успешно сняты
          content:
            text/plain:
              schema:
                type: string
                example: "Funds have been withdrawn: -500.00"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /user/card/transfer:
    put:
      summary: Перевод между картами
      description: Переводит средства с одной карты на другую
      operationId: transfer
      tags:
        - User Card Management
      parameters:
        - name: fromId
          in: query
          description: ID карты отправителя
          required: true
          schema:
            type: integer
            format: int64
            example: 12345
        - name: toId
          in: query
          description: ID карты получателя
          required: true
          schema:
            type: integer
            format: int64
            example: 67890
        - name: amount
          in: query
          description: Сумма перевода
          required: true
          schema:
            type: number
            format: decimal
            minimum: 0.01
            example: 300.00
      responses:
        '200':
          description: Перевод выполнен успешно
          content:
            text/plain:
              schema:
                type: string
                example: "Funds have been transferred\nCard-id:12345 -300.00\nCard-id:67890 +300.00"
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    CardDTO:
      type: object
      required:
        - id
        - cardNumber
        - ownerId
        - expiryDate
        - status
        - balance
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор карты
          example: 12345
        cardNumber:
          type: string
          description: Номер карты (16 цифр)
          pattern: '^\d{16}$'
          example: "1234567812345678"
        ownerId:
          type: string
          format: uuid
          description: ID владельца карты
          example: "123e4567-e89b-12d3-a456-426614174000"
        expiryDate:
          type: string
          format: year-month
          description: Дата окончания действия в формате ГГГГ-ММ
          pattern: '^\d{4}-\d{2}$'
          example: "2025-12"
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: bigdecimal
          description: Баланс карты
          minimum: 0
          example: 1500.75
...
    CardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - EXPIRED
      description: Статусы карты
      example: "ACTIVE"
...
    UserCreationDTO:
      type: object
      properties:
        username:
          type: string
          description: Имя пользователя
          minLength: 3
          maxLength: 50
        password:
          type: string
          description: Пароль пользователя
          minLength: 6
          maxLength: 100
          format: password
      required:
        - username
        - password
...
    Error:
      type: object
      required:
        - status
        - error
        - message
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2024-01-15T10:30:00Z"
        status:
          type: integer
          description: HTTP статус код
        error:
          type: string
          description: Тип ошибки
        message:
          type: string
          description: Сообщение об ошибке
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен аутентификации в формате Bearer token
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []